--- frontend/__tests__/unit/NetworkHeader.snapshot.jsx
@@ -1,8 +1,0 @@
-import { render } from "@testing-library/react";
-import NetworkHeader from "components/NetworkHeader";
-import { testNetwork } from "./NetworkHeader.test";
-
-it("renders HomeLoggedOut unchanged", () => {
-  const { container } = render(<NetworkHeader network={testNetwork} />);
-  expect(container).toMatchSnapshot();
-});
